Fixing umount delete test failure
After the longhorn refactor, this case was failing as Linux throws error
when the test case was trying to unmount the mount path when an active
volume has a replica running and storing meta files in this mount path.
The following code changes fixes it using -l flag to accomplish the
motive that the test case was trying to achieve.