Add backup status test for unavailable replicas
This test is used to make sure that we do not try to get the backup
status from unavailable replicas. This could lead to the engine backup
status retrieval process exceeding the maximum allowed time, after which
the manager would terminate that process. Since we wouldn't be able to
retrieve updated backup statuses we could no longer show new backups in
the frontend.

Longhorn #1230

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>