Add skeleton for backup block deletion test
We want to make sure that we only delete non referenced backup blocks,
we also don't want to delete blocks while there other backups in progress.
The reason for this is that we don't yet know which blocks are required by
the in progress backup, so blocks deletion could lead to a faulty backup.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>