Fix test_salvage_auto_crash_all_replicas test case
It is hard to capture the pending state of workload pod.
Moreover, we only need to make sure that the pod is in
"running" state and wait_for_pod_remount() return true.
That is enough to conclude that the pod was restarted
and the volume was remount. Therefore, we removed the
wait_for_pod_phase(core_api, pod_name, pod_phase="Pending")

Longhorn #1914

Signed-off-by: Phan Le <phan.le@rancher.com>