Fix test_backup_status_for_unavailable_replicas
The behavior changes after longhorn 3035.

Previous, the volume's backupStatus reflects the engine's backup status directly.
So, even the backup is completed, if we delete the corresponding replica, the
volume's backupStatus shows the engine's backup status with an error message.

Now, after longhorn 3035 was introduced, the volume's backupStatus reflects
the Backup CR status, not the engine's backup status anymore.
So, after the backup is completed, even we delete the corresponding replica,
the volume's backupStatus shows the Backup state is Completed.

We change the test case to during the backup creation in progress,
deletes the responsible replica to make the backup falls into Error
state and with an error message.

Longhorn 3035

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>