Print pytest logs in jenkins console output
To fetch the test report, we have to "kubectl cp" the report,
but there is no way to "kubectl cp" a completed pod, so another
keepalive pod is created and we can get the test report from it.

After resolved the "kubectl cp" issue above, we don't have to
use "kubectl logs" to print out the test report anymore. We can
use "kubectl logs" to print pytest logs normally.

Signed-off-by: Yang Chiu <yang.chiu@suse.com>