test: fix flaky test case test_cloning_with_detached_source_volume
by deferring the snapshot creation check.

the source volume will be automatically attached and detached during
the clone process, since it will be detached immediatley after the clone
is completed, it's nearly impossible to always catch the moment of the
source volume is still attached and get the snapshot list.
defer the snapshot creation check to the end of the test case by manually
attaching the source volume.

Signed-off-by: Yang Chiu <yang.chiu@suse.com>