test: add try/except for generate support bundle function to avoid internal error
In gke environment, the credential will keep expire and automatically renew during
the test running. If the test tries to make requests when the credntial expired and
without this try/except, it would cause internal error and fail the whole test process
instead of just failing the current test case.

Signed-off-by: Yang Chiu <yang.chiu@suse.com>