ci: refactor to have separate gke pipeline
To be able to run gke test, we need to install gcloud in a test build,
but it causes each test build occupies significant amount of disk space
(> 4GB) and causes infra node disk pressure and affects the services
if lots of test builds are triggered at the same time.
So refactor to separate gke pipeline first to avoid installing gcloud
in every build.

Signed-off-by: Yang Chiu <yang.chiu@suse.com>