test: fix flaky test case test_orphan_with_same_orphaned_dir_name_in_another_disk
by retry orphan delete api.

The client.delete(orphan) function is somehow not reliable. It's possible we called
this api but it takes no effect and the orphan still exists. Also the api doesn't
return anything so there's no way to know it succeeds or not by checking the response.

Add retry mechanism to ensure the orphan can be deleted as expected.

Signed-off-by: Yang Chiu <yang.chiu@suse.com>