tests: Add object endpoint test
Add first tests and fixtures for the object endpoint controller.

There are two fixtures for testing the object endpoint controller:
A storage class and an object endpoint.
Although there is nothing special about the storage class, it is there
as a fixture since it needs to be ensured a known storage class exists
to be able to create an object endpoint.
The object endpoint exists as fixture, because this makes creation and
deletion easier, ensuring that even in the event of a failed test the
object endpoint is cleaned up.

In the object endpoint creation test, several assertions are made on
the resources connected to the object endpoint to ensure creation was
successful and the resulting object endpoint works as expected.

In the object endpoint deletion test, an object endpoint is created, but
then it is explicitly deleted. After that, several assertions are made
to ensure no resources are left behind.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>